+++ client/gldt_swap_app/src/components/ui/Select.tsx
@@ -41,7 +41,7 @@ const Select = ({
               </div>
             </div>
             <span className="pointer-events-none absolute inset-y-0 right-0 flex items-center pr-2">
-              <ChevronUpDownIcon className="h-5 w-5" aria-hidden="true" />
+              <ChevronUpDownIcon className="h-5 w-5" />
             </span>
           </ListboxButton>